[fix/#19] AuctionBid 패키지 분리
