--- src/main/java/com/api/jaebichuri/auction/entity/AuctionBid.java
@@ -1,34 +1,0 @@
-package com.api.jaebichuri.auction.entity;
-
-import com.api.jaebichuri.member.entity.Member;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AuctionBid {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long bidPrice;
-
-    @Column(nullable = false)
-    private LocalDateTime bidTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "auction_id", nullable = false)
-    private Auction auction;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member bidder;
-
-}
